LU-19059 lov: fix lov_stripe_set
Prevent overflow on multiplication of 32 bit integers.

Test to illustrate the issue is added to hit overflow in both branches
of lov_stripe_set().

Fixes: 6eee4ea5b6 ("LU-6174 lov: use standard Linux 64 divison macros")
HPE-bug-id: LUS-12782
Reviewed-on: https://review.whamcloud.com/c/fs/lustre-release/+/59436
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/lov/lov_offset.c
@@ -35,7 +35,7 @@ u64 stripe_width(struct lov_stripe_md *lsm, unsigned int index)
 u64 lov_stripe_size(struct lov_stripe_md *lsm, int index, u64 ost_size,
 		    int stripeno)
 {
-	u32 ssize = lsm->lsm_entries[index]->lsme_stripe_size;
+	u64 ssize = lsm->lsm_entries[index]->lsme_stripe_size;
 	u32 stripe_size;
 	u64 swidth;
 	u64 lov_size;
